refactor(phase8): extract payload/bind/normalization helpers to codegen/payload_tokens.rs
Moves CreatePayloadField, create_payload_fields, bind_field_value_tokens,
integer_to_i64_tokens, update_payload_fields, normalization_tokens, and
all validation/normalization helpers (~270 lines) into their own submodule.

codegen.rs: 711 → 450 lines (−37 %).  Original 4,336 → 450 (−90 %).
All 233 tests pass.
